# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Social Networks API
  version: 1.0.0
  description: |
    API REST du nouveau service d'événements et de groupes de Facebook (TP Node.js / Express / MongoDB).

    **Authentification** : `POST /auth/register` ou `POST /auth/login` renvoient un token JWT (valable 24h)
    à envoyer dans le header `Authorization: Bearer <token>`.

    **Format des erreurs** : toutes les erreurs ont la forme `{ code, message, errors }`, où `errors`
    est la liste des problèmes à corriger (ex : erreurs de validation).

    **Listes** : les listes sont paginées avec `?page=1&limit=20` (100 maximum) et renvoient
    `{ data, pagination: { page, limit, total, pages } }`.

    **Visibilité** :
    - un groupe *secret* ou un événement *privé* répond `404` à ceux qui n'en font pas partie (il « n'existe pas » pour eux) ;
    - un groupe *privé* est visible (nom, description) mais ses membres, son fil et ses événements sont réservés aux membres (`403`).
  contact:
    name: Saad Lagzouli
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Développement
...
security:
  - bearerAuth: []
paths:
...
  /threads/{id}:
    get:
      operationId: getThread
      tags: [Threads]
      summary: Détail d'un fil de discussion
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Fil avec son groupe ou son événement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Thread' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /threads/{id}/messages:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getMessages
      tags: [Threads]
      summary: Messages du fil (ordre chronologique)
      description: Lecture = membres du groupe (ou tout le monde si groupe public) / participants de l'événement.
      parameters:
        - { name: reply_to, in: query, description: Uniquement les réponses à ce message, schema: { type: string } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageList' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createMessage
      tags: [Threads]
      summary: Publier un message ou répondre à un message
      description: |
        Membres du groupe / participants de l'événement.
        Si le groupe n'autorise pas les membres à publier (`allow_member_posts: false`), seuls les administrateurs
        publient de nouveaux messages, mais tous les membres peuvent répondre (`reply_to`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MessageInput' }
      responses:
        '201':
          description: Message créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /threads/{id}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - { name: messageId, in: path, required: true, schema: { type: string } }
    patch:
      operationId: updateMessage
      tags: [Threads]
      summary: Modifier son message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, maxLength: 2000 }
      responses:
        '200': { description: Message modifié }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteMessage
      tags: [Threads]
      summary: Supprimer un message et ses réponses
      description: Auteur, administrateur du groupe ou organisateur de l'événement.
      responses:
        '200': { description: 'Message supprimé (avec `deleted_replies`)' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------------ Albums
...
  /events/{id}/polls:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getPolls
      tags: [Polls]
      summary: Sondages de l'événement (participant)
      description: Chaque sondage indique `answered` (ai-je déjà répondu ?).
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PollList' }
    post:
      operationId: createPoll
      tags: [Polls]
      summary: Créer un sondage (organisateur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PollInput' }
      responses:
        '201':
          description: Sondage créé (chaque question et chaque réponse reçoit un `_id`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Poll' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /polls/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getPollById
      tags: [Polls]
      summary: Détail d'un sondage avec mes réponses
      responses:
        '200':
          description: Sondage avec `my_answers` (null si pas encore répondu)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Poll' }
    delete:
      operationId: deletePoll
      tags: [Polls]
      summary: Supprimer un sondage et ses réponses (organisateur)
      responses:
        '200': { description: Sondage supprimé }
        '403': { $ref: '#/components/responses/Forbidden' }
  /polls/{id}/answers:
    post:
      operationId: answerPoll
      tags: [Polls]
      summary: Répondre au sondage (participant)
      description: Une seule participation par personne (`409`). Il faut répondre à chaque question avec un seul choix.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PollAnswerInput' }
      responses:
        '201': { description: Réponses enregistrées }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /polls/{id}/results:
    get:
      operationId: getResults
      tags: [Polls]
      summary: Résultats du sondage (participant)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Nombre de votes par réponse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PollResults' }

  # ------------------------------------------------------------------ Ticketing
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, example: 6702f1c2a1b2c3d4e5f60718 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 400, message: Bad Request, errors: ['end_date doit être après start_date'] }
...
    Forbidden:
      description: Action non autorisée pour cet utilisateur
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 403, message: Forbidden, errors: ["action réservée aux organisateurs de l'événement"] }
    NotFound:
      description: Ressource introuvable (ou invisible pour cet utilisateur)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 404, message: Event Not Found }
    Conflict:
      description: Conflit avec l'état actuel (doublon, plus de place, règle métier)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 409, message: Conflict, errors: ["une personne ne peut obtenir qu'un seul billet par événement"] }
  schemas:
    Error:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        errors: { type: array, items: { type: string } }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        pages: { type: integer, example: 3 }
    ObjectId:
      type: string
      example: 6702f1c2a1b2c3d4e5f60718
    Timestamps:
      type: object
      properties:
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ---------- users
    UserPublic:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        firstname: { type: string, example: Saad }
        lastname: { type: string, example: Lagzouli }
        avatar: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
...
    Thread:
      type: object
      description: Lié à un groupe OU à un événement, jamais aux deux
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        group: { type: object, nullable: true, description: Groupe du fil (ou null) }
        event: { type: object, nullable: true, description: Événement du fil (ou null) }
        messages_count: { type: integer }
        created_at: { type: string, format: date-time }
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string, maxLength: 2000, example: Qui vient samedi ? }
        reply_to: { type: string, nullable: true, description: Message auquel on répond }
    Message:
      allOf:
        - $ref: '#/components/schemas/MessageInput'
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            thread: { $ref: '#/components/schemas/ObjectId' }
            author: { $ref: '#/components/schemas/UserPublic' }
    MessageList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Message' } }
        pagination: { $ref: '#/components/schemas/Pagination' }

    # ---------- albums
...
    PollInput:
      type: object
      required: [title, questions]
      properties:
        title: { type: string, maxLength: 200, example: Organisation du week-end }
        questions:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: object
            required: [label, choices]
            properties:
              label: { type: string, maxLength: 300, example: Quel repas ? }
              choices:
                type: array
                minItems: 2
                maxItems: 10
                items: { type: string }
                example: [Pique-nique, Restaurant]
    Poll:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        event: { $ref: '#/components/schemas/ObjectId' }
        title: { type: string }
        questions:
          type: array
          items:
            type: object
            properties:
              _id: { $ref: '#/components/schemas/ObjectId' }
              label: { type: string }
              choices:
                type: array
                items:
                  type: object
                  properties:
                    _id: { $ref: '#/components/schemas/ObjectId' }
                    label: { type: string }
        answered: { type: boolean, description: Dans la liste uniquement }
        my_answers: { type: array, nullable: true, items: { $ref: '#/components/schemas/Answer' }, description: Dans le détail uniquement }
        created_by: { $ref: '#/components/schemas/ObjectId' }
        created_at: { type: string, format: date-time }
    PollList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Poll' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    Answer:
      type: object
      required: [question, choice]
      properties:
        question: { $ref: '#/components/schemas/ObjectId' }
        choice: { $ref: '#/components/schemas/ObjectId' }
    PollAnswerInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          description: Une réponse pour chaque question du sondage
          items: { $ref: '#/components/schemas/Answer' }
    PollResults:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        title: { type: string }
        respondents: { type: integer }
        questions:
          type: array
          items:
            type: object
            properties:
              _id: { $ref: '#/components/schemas/ObjectId' }
              label: { type: string }
              choices:
                type: array
                items:
                  type: object
                  properties:
                    _id: { $ref: '#/components/schemas/ObjectId' }
                    label: { type: string }
                    votes: { type: integer }

    # ---------- ticketing
...